Pulling shared functionality to BaseApp class
